+++ test/datasource/mock/remote/test_put_datasource.dart
@@ -18,10 +18,10 @@ final class TestPutDataSource
   Map<String, String> get pathModification => {':id': '123'};
 
   @override
-  List<int> get admissibleStatusCode => [200];
+  Set<int> get admissibleStatusCode => {200};
 
   @override
-  List<int> get inadmissibleStatusCode => [404];
+  Set<int> get inadmissibleStatusCode => {404};
 
   @override
   PutParams generateCallRequirement({required Params params}) {